split overlay layer and add all input device
I split the overlay layer into an explicit locking layer and an
explicit input area layer.
This avoids ambiguous z-order for applications that should have a well
defined order.
I also added intended usecases to the documentation. Those are mainly
to set up for discussion.
It would be beneficial if everyone has a simliar model of how this
protocol should be used, so these should be informational.

I also added an "all" flag to input devices. This is intended for
session locking tools, to be able to lock all input, without having to
be updated for the mos current protocol spec with newer input devices.