backend/wayland: don't destroy wl_buffer on release
When the parent compositor sends a wl_buffer.release event, we need to
unlock the wlr_buffer to notify the child compositor that it can be
re-used. However we don't need to destroy the wl_buffer. We can keep it
around in case the child compositor re-uses and re-submits the
wlr_buffer.