backend: build: all to be an optional module
I think the session and multi are the necessary modules,
since lacks of alternative allocator or renderer, this
work is be done yet.

libinput:
A few usage doesn't need input at all, for example those
monitors in the elevator room.

x11:
change library provider message

headless:
We just need an alloactor here, gbm would be not the only
option.

wayland:
We just need an alloactor here, the gbm may not be the only
option for the nested wayland.

Signed-off-by: ayaka <ayaka@soulik.info>